+++ SanguineGenesis/GameLogic/Data/Entities/Unit.cs
@@ -25,7 +25,7 @@ namespace SanguineGenesis.GameLogic.Data.Entities
         /// </summary>
         public override float Range { get; }
 
-        public Unit(Faction faction, string unitType, decimal maxHealth, float viewRange, decimal maxEnergy, List<Ability> abilities, Vector2 position, float range, bool physical)
+        public Unit(Faction faction, string unitType, float maxHealth, float viewRange, float maxEnergy, List<Ability> abilities, Vector2 position, float range, bool physical)
             :base(faction, unitType, maxHealth, viewRange, maxEnergy, physical, abilities)
         {
             Position = position;
